pyproject.toml
[project]
name = "portfolio-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.7.0",
    "flask>=3.1.1",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.1.0",
]
